+++ lib/themes/styles.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+
+// For a non-Cupertino version of this app
+ThemeData appTheme() {
+  return ThemeData(
+    primaryColor: Colors.white,
+    primarySwatch: Colors.blue,
+  );
+}
